backend/requirements.txt
fastapi
uvicorn
python-gnupg
pydantic
python-multipart
